handle update passcode flow

+++ src/screen/confirmPasscode.js
@@ -10,7 +10,7 @@ import Keyboard from '../component/keyboard'
 import Screen from '../component/screenComponent'
 import { HeadSpace } from '../component/headSpace'
 import { navigateAction, navigateReset } from '../redux/actions'
-import { requestRegister } from '../redux/actions/root-active'
+import { confirmPasscode } from '../redux/actions/root-active'
 import { root } from '../redux/actions/commonAction'
 import getnativeModules from '../containers/hoc/infoAppNativeModules'
 
@@ -19,11 +19,19 @@ const defaultPasscode = {
   number: '',
 }
 
+const handleLinkPage = currFlow => {
+  switch (currFlow) {
+    case 'updatePasscode': return 'portSuggestion'
+    case 'forgetPasscode': return 'user/forgot-password/reset-passcode'
+    default: return 'condi'
+  }
+}
+
 const mapToProps = ({ root, passcode }) => ({ root, passcode })
 const dispatchToProps = dispatch => ({
   navigateAction: bindActionCreators(navigateAction, dispatch),
   navigateReset: bindActionCreators(navigateReset, dispatch),
-  requestRegister: bindActionCreators(requestRegister, dispatch),
+  confirmPasscode: bindActionCreators(confirmPasscode, dispatch),
   updateRoot: bindActionCreators(root, dispatch)
 })
 @connect(mapToProps, dispatchToProps)
@@ -48,10 +56,10 @@ export default class extends React.Component {
     }
     if (obj.number.length === 6) {
       if (this.props.passcode.passcode === data.password) {
-        this.props.requestRegister(data, this.props.root.access_token)
+        this.props.confirmPasscode(data, { token: this.props.root.access_token, currFlowUP: this.props.root.currFlowUP,})
           .then(res => {
             if (res.success) {
-              this.props.navigateAction({ ...this.props, page: 'condi' })
+              this.props.navigateAction({ ...this.props, page: handleLinkPage(this.props.root.currFlowUP) })
               AsyncStorage.setItem('access_token', res.result.access_token)
               AsyncStorage.setItem('user_token', res.result.user_token)
               NativeModules.KMyFundOnboarding.saveRegisterFlag(NativeModules.KMyFundOnboarding.STATUS_NEW_CUSTOMER)
